Rename 'GitHubPage.tsx' to 'tldrPageUrl.tsx'
